# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lab Booking System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookingBookerType =
  (typeof BookingBookerType)[keyof typeof BookingBookerType];

export const BookingBookerType = {
  faculty: "faculty",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: string;
  bookerName: string;
  bookerType: BookingBookerType;
  purpose: string;
  labName: string;
  date: string;
  startTime: string;
  endTime: string;
  studentCount: number;
  status: BookingStatus;
  rejectionReason?: string | null;
  createdAt: string;
}

export type CreateBookingBodyBookerType =
  (typeof CreateBookingBodyBookerType)[keyof typeof CreateBookingBodyBookerType];

export const CreateBookingBodyBookerType = {
  faculty: "faculty",
} as const;

export interface CreateBookingBody {
  idToken: string;
  bookerName: string;
  bookerType: CreateBookingBodyBookerType;
  purpose: string;
  labName: string;
  date: string;
  startTime: string;
  endTime: string;
  studentCount: number;
}

export interface RejectBookingBody {
  reason?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LabSchedule {
  labName: string;
  bookings: Booking[];
}

export interface DaySchedule {
  date: string;
  labs: LabSchedule[];
}

export type BookingStatsLabBreakdownItem = {
  labName: string;
  count: number;
};

export interface BookingStats {
  totalBookings: number;
  pendingApprovals: number;
  todayBookings: number;
  labBreakdown: BookingStatsLabBreakdownItem[];
}

export type FacilityType = (typeof FacilityType)[keyof typeof FacilityType];

export const FacilityType = {
  lab: "lab",
  classroom: "classroom",
} as const;

export type FacilityStatus =
  (typeof FacilityStatus)[keyof typeof FacilityStatus];

export const FacilityStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Facility {
  id: string;
  name: string;
  displayName: string;
  type: FacilityType;
  capacity: number;
  systemCount?: number | null;
  seatCount?: number | null;
  description?: string | null;
  status: FacilityStatus;
}

export type CreateFacilityBodyType =
  (typeof CreateFacilityBodyType)[keyof typeof CreateFacilityBodyType];

export const CreateFacilityBodyType = {
  lab: "lab",
  classroom: "classroom",
} as const;

export type CreateFacilityBodyStatus =
  (typeof CreateFacilityBodyStatus)[keyof typeof CreateFacilityBodyStatus];

export const CreateFacilityBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateFacilityBody {
  name: string;
  displayName: string;
  type: CreateFacilityBodyType;
  capacity: number;
  systemCount?: number | null;
  seatCount?: number | null;
  description?: string | null;
  status: CreateFacilityBodyStatus;
}

export type ListBookingsParams = {
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
  /**
   * Filter by lab name
   */
  labName?: string;
};

export type AdminListBookingsParams = {
  /**
   * Filter by status
   */
  status?: AdminListBookingsStatus;
};

export type AdminListBookingsStatus =
  (typeof AdminListBookingsStatus)[keyof typeof AdminListBookingsStatus];

export const AdminListBookingsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  cancelled: "cancelled",
} as const;

export type GetScheduleParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
};