# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lab Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingBookerType } from "./bookingBookerType";
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: string;
  bookerName: string;
  bookerType: BookingBookerType;
  purpose: string;
  labName: string;
  date: string;
  startTime: string;
  endTime: string;
  studentCount: number;
  status: BookingStatus;
  rejectionReason?: string | null;
  createdAt: string;
}