# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lab Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingBodyBookerType } from "./createBookingBodyBookerType";

export interface CreateBookingBody {
  idToken: string;
  bookerName: string;
  bookerType: CreateBookingBodyBookerType;
  purpose: string;
  labName: string;
  date: string;
  startTime: string;
  endTime: string;
  studentCount: number;
}