# lib/api-zod/src/generated/types/createFacilityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lab Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFacilityBodyStatus } from "./createFacilityBodyStatus";
import type { CreateFacilityBodyType } from "./createFacilityBodyType";

export interface CreateFacilityBody {
  name: string;
  displayName: string;
  type: CreateFacilityBodyType;
  capacity: number;
  systemCount?: number | null;
  seatCount?: number | null;
  description?: string | null;
  status: CreateFacilityBodyStatus;
}