# lib/api-zod/src/generated/types/facility.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lab Booking System API
 * OpenAPI spec version: 0.1.0
 */
import type { FacilityStatus } from "./facilityStatus";
import type { FacilityType } from "./facilityType";

export interface Facility {
  id: string;
  name: string;
  displayName: string;
  type: FacilityType;
  capacity: number;
  systemCount?: number | null;
  seatCount?: number | null;
  description?: string | null;
  status: FacilityStatus;
}